// types/PokemonDetail.ts
export interface Ability {
  ability: Ability2;
  is_hidden: boolean;
  slot: number;
}

export interface Ability2 {
  name: string;
  url: string;
}
// ...
export interface PastAbility {
  abilities: Ability3[];
  generation: Generation;
}

export interface Ability3 {
  ability: Ability2;
  is_hidden: boolean;
  slot: number;
}

export interface Generation {
  name: string;
  url: string;
}